Revert unnecesary change on IERC721Mock

+++ contracts/interfaces/IERC721Mock.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts-v4/token/ERC721/IERC721.sol";
 
 interface IERC721Mock is IERC721 {
-  function safeMint(address to, uint256 tokenId) external;
+  function safeMint(address to, uint tokenId) external;
 
-  function isApprovedOrOwner(address spender, uint256 tokenId) external returns (bool);
+  function isApprovedOrOwner(address spender, uint tokenId) external returns (bool);
 }
